Avoid duplicate exports
https://github.com/juju-solutions/interface-mount/pull/4 allows for multiple
mount requests to share the same export name.  This could previously have
caused this charm to write an exports file with duplicate lines, which in turn
caused `exportfs` to issue a "duplicated export entries" error message and exit
with status 1.

`nfs_relation_changed` now deduplicates the lines it writes to its exports
file.  It still needs to send an appropriate response to each `mount` relation,
so it now keeps track of those responses separately.